random: move to libtrx
