+++ src/libtrx/game/random.c
@@ -1,11 +1,20 @@
 #include "game/random.h"
 
-#include "game/phase/phase.h"
+#include "log.h"
 
-#include <libtrx/log.h>
+#include <time.h>
 
 static int32_t m_RandControl = 0xD371F947;
 static int32_t m_RandDraw = 0xD371F947;
+static bool m_IsDrawFrozen = false;
+
+void Random_Seed(void)
+{
+    time_t lt = time(0);
+    struct tm *tptr = localtime(&lt);
+    Random_SeedControl(tptr->tm_sec + 57 * tptr->tm_min + 3543 * tptr->tm_hour);
+    Random_SeedDraw(tptr->tm_sec + 43 * tptr->tm_min + 3477 * tptr->tm_hour);
+}
 
 void Random_SeedControl(int32_t seed)
 {
@@ -30,9 +39,13 @@ int32_t Random_GetDraw(void)
     // Allow draw RNG to advance only during initial game setup (for such things
     // as caustic initialisation) and normal game play. RNG should remain static
     // when the game output is paused e.g. inventory, pause screen etc.
-    PHASE_ENUM phase = Phase_Get();
-    if (phase == PHASE_NULL || phase == PHASE_GAME) {
+    if (!m_IsDrawFrozen) {
         m_RandDraw = 0x41C64E6D * m_RandDraw + 0x3039;
     }
     return (m_RandDraw >> 10) & 0x7FFF;
+}
+
+void Random_FreezeDraw(bool is_frozen)
+{
+    m_IsDrawFrozen = is_frozen;
 }
