+++ src/tr1/game/random.h
@@ -1,8 +1,3 @@
 #pragma once
 
-#include <stdint.h>
-
-void Random_SeedControl(int32_t seed);
-void Random_SeedDraw(int32_t seed);
-int32_t Random_GetControl(void);
-int32_t Random_GetDraw(void);
+#include <libtrx/game/random.h>
