+++ src/tr2/game/random.h
@@ -1,9 +1,3 @@
 #pragma once
 
-#include <stdint.h>
-
-void Random_SeedControl(int32_t seed);
-void Random_SeedDraw(int32_t seed);
-int32_t Random_GetControl(void);
-int32_t Random_GetDraw(void);
-void Random_Seed(void);
+#include <libtrx/game/random.h>
