+++ src/tr2/meson.build
@@ -255,7 +255,6 @@ sources = [
   'game/phase/phase_game.c',
   'game/phase/phase_inventory.c',
   'game/phase/phase_stats.c',
-  'game/random.c',
   'game/render/common.c',
   'game/render/hwr.c',
   'game/render/priv.c',
